variables fixed, weird error found

+++ selections.h
@@ -34,7 +34,7 @@ bool trig_num(const char* final_state){
 }
 
 
-int isMuon(Int_t num, vector<float>* muPt, vector<float>* muEta, vector<unsigned short>* muIDbit, vector<float>* muDz, vector<float>* muD0, vector<float>* muPFNeuIso, vector<float>* muPFPhoIso, vector<float>* muPFPUIso, vector<float>* muPFChIso){
+int isMuon(Int_t num, vector<float>* muPt, vector<float>* muEta, vector<int>* muIDbit, vector<float>* muDz, vector<float>* muD0, vector<float>* muPFNeuIso, vector<float>* muPFPhoIso, vector<float>* muPFPUIso, vector<float>* muPFChIso){
     int number = -1;
     int extra = 0;
     for (int counter=0; counter<num; counter++){
@@ -90,7 +90,7 @@ int isTau(Int_t num, vector<float>* tauPt, vector<float>* tauEta, vector<int>* t
 
 
 
-bool rejectMuon(Int_t num, vector<float>* muPt, vector<float>* muEta, vector<float>* muD0, vector<float>* muDz, vector<unsigned short>* muIDbit, vector<float>* muPFNeuIso, vector<float>* muPFPhoIso, vector<float>* muPFPUIso, vector<float>* muPFChIso){
+bool rejectMuon(Int_t num, vector<float>* muPt, vector<float>* muEta, vector<float>* muD0, vector<float>* muDz, vector<int>* muIDbit, vector<float>* muPFNeuIso, vector<float>* muPFPhoIso, vector<float>* muPFPUIso, vector<float>* muPFChIso){
     bool reject=false;
     for (int counter=0; counter<num; counter++){
         float thisiszero=0;
